Autoconfig: Optimize autoconfig
This commit uses the inbuilt primitives for caching, and not requesting things as often to decrease the
lag when dealing with atomic server config changes.

Previously, it could take up to 7 seconds to do nothing, as we were uselessly putting unchanged things, and
repeatedly requesting the same thing.

Now the system takes less than a second.